Added bubblesort function with custom comparison function
To allow the use of templated functions for the sorting, the function must be
implemented in the header so that the compiler can see how to generate code for
different template arguments. This is annoying, but there doesn't seem to be any
other way of doing it which makes sense given that this is supposed to be a
library. Doxygen stuff probably needs to remain in the .cpp file, but maybe it
doesn't make much difference just putting it in the .hpp.